chore: upgrade to libsignal 0.77.1 (#361)
* Update protos
* Restore relinking
* Bump MSRV to 1.83
* ::No
* Please clippy

---------

+++ src/push_service/account.rs
@@ -47,7 +47,7 @@ pub struct DeviceId {
 #[derive(Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct DeviceInfo {
-    pub id: i64,
+    pub id: u8,
     pub name: Option<String>,
     #[serde(with = "chrono::serde::ts_milliseconds")]
     pub created: DateTime<Utc>,
